moved all caches to a dedicated folder for cleaner code. renamed authentication cache since it's not a cache. fixed imports broken by better proto i missed earlier
