+++ galaxy-integration-steam.pyproj
@@ -76,12 +76,13 @@
     <Compile Include="src\persistent_cache_state.py" />
     <Compile Include="src\plugin.py" />
     <Compile Include="src\registry_monitor.py" />
-    <Compile Include="src\steam_network\authentication_cache.py" />
-    <Compile Include="src\steam_network\cache_proto.py" />
+    <Compile Include="src\steam_network\authentication_data.py" />
+    <Compile Include="src\steam_network\caches\cache_proto.py" />
+    <Compile Include="src\steam_network\caches\__init__.py" />
     <Compile Include="src\steam_network\enums.py" />
-    <Compile Include="src\steam_network\friends_cache.py" />
-    <Compile Include="src\steam_network\games_cache.py" />
-    <Compile Include="src\steam_network\local_machine_cache.py" />
+    <Compile Include="src\steam_network\caches\friends_cache.py" />
+    <Compile Include="src\steam_network\caches\games_cache.py" />
+    <Compile Include="src\steam_network\caches\local_machine_cache.py" />
     <Compile Include="src\steam_network\presence.py" />
     <Compile Include="src\steam_network\protocol\consts.py" />
     <Compile Include="src\steam_network\protocol\job.py" />
@@ -103,23 +104,21 @@
     <Compile Include="src\steam_network\protocol\messages\steammessages_webui_friends.py" />
     <Compile Include="src\steam_network\protocol\messages\__init__.py" />
     <Compile Include="src\steam_network\protocol\protobuf_client.py" />
-    <Compile Include="src\steam_network\protocol\protobuf_parser.py" />
-    <Compile Include="src\steam_network\protocol\rsa_message.py" />
     <Compile Include="src\steam_network\protocol\steam_types.py" />
     <Compile Include="src\steam_network\protocol\__init__.py" />
     <Compile Include="src\steam_network\protocol_client.py" />
-    <Compile Include="src\steam_network\stats_cache.py" />
+    <Compile Include="src\steam_network\caches\stats_cache.py" />
     <Compile Include="src\steam_network\steam_auth_polling_data.py" />
     <Compile Include="src\steam_network\steam_http_client.py" />
     <Compile Include="src\steam_network\steam_network_controller.py" />
     <Compile Include="src\steam_network\steam_network_model.py" />
     <Compile Include="src\steam_network\steam_network_view.py" />
     <Compile Include="src\steam_network\steam_public_key.py" />
-    <Compile Include="src\steam_network\times_cache.py" />
-    <Compile Include="src\steam_network\user_info_cache.py" />
+    <Compile Include="src\steam_network\caches\times_cache.py" />
+    <Compile Include="src\steam_network\caches\user_info_cache.py" />
     <Compile Include="src\steam_network\utils.py" />
     <Compile Include="src\steam_network\w3_hack.py" />
-    <Compile Include="src\steam_network\websocket_cache_persistence.py" />
+    <Compile Include="src\steam_network\caches\websocket_cache_persistence.py" />
     <Compile Include="src\steam_network\websocket_client.py" />
     <Compile Include="src\steam_network\websocket_list.py" />
     <Compile Include="src\steam_network\__init__.py" />
@@ -177,6 +176,7 @@
     <Folder Include="src\steam_network" />
     <Folder Include="src\steam_network\custom_login" />
     <Folder Include="src\steam_network\custom_login\css" />
+    <Folder Include="src\steam_network\caches\" />
     <Folder Include="src\steam_network\protocol" />
     <Folder Include="src\steam_network\protocol\messages\" />
     <Folder Include="tests" />
