+++ src/steam_network/caches/stats_cache.py
